+++ @types/index.d.ts
@@ -329,6 +329,7 @@ declare module 'opentok-react-native' {
       canForceMute: boolean;
       canPublish: boolean;
       canSubscribe: boolean;
+      canForceDisconnect: boolean;
     }>;
 
     /**
@@ -340,6 +341,11 @@ declare module 'opentok-react-native' {
      * Mutes a stream in the session.
      */
     forceMuteStream: (streamId: string) => Promise<void>;
+
+    /**
+     * Forces a client to disconnect from the session.
+     */
+    forceDisconnect: (connectionId: string) => Promise<void>;
 
     /**
      * Disables the force mute state for the session.
@@ -459,6 +465,31 @@ declare module 'opentok-react-native' {
      * To publish a screen-sharing stream, set this property to "screen". If you do not specify a value, this will default to "camera".
      */
     videoSource?: VideoSource;
+
+    /**
+     * The video content hint fpr the publisher (either "none", "motion", "detail", or "text").
+     */
+    videoContentHint?: string;
+
+    /**
+     * Whether to allow use of scalable video for a publisher that has the videoSource set "screen".
+     */
+    scalableScreenshare?: boolean;
+
+    /**
+     * If set to false, the microphone will be automatically switched off when the publish has muted.
+     */
+    allowAudioCaptureWhileMuted?: boolean;
+
+    /**
+     * The maximum video bitrate for the published stream (between 5,000 and 10,000,000).
+     */
+    maxVideoBitrate?: number;
+
+    /**
+     * The video bitrate preset to use for the published stream. Ignored if maxVideoBitrate is set.
+     */
+    videoBitratePreset?: 'default' | 'bw_saver' | 'extra_bw_saver';
   }
 
   interface OTPublisherEventHandlers {