+++ android/src/main/java/com/opentokreactnative/OTRNPublisherManager.kt
@@ -107,6 +107,16 @@ class OTRNPublisherManager(context: ReactApplicationContext) :
         view.setVideoContentHint(value)
     }
 
+    @ReactProp(name = "maxVideoBitrate")
+    override public fun setMaxVideoBitrate(view: OTRNPublisher, value: Int) {
+        view.setMaxVideoBitrate(value)
+    }
+
+    @ReactProp(name = "videoBitratePreset")
+    override public fun setVideoBitratePreset(view: OTRNPublisher, value: String?) {
+        view.setVideoBitratePreset(value)
+    }
+
     @ReactProp(name = "cameraPosition")
     override public fun setCameraPosition(view: OTRNPublisher, value: String?) {
         view.setCameraPosition(value)
@@ -145,6 +155,11 @@ class OTRNPublisherManager(context: ReactApplicationContext) :
     @ReactProp(name = "scalableScreenshare")
     override public fun setScalableScreenshare(view: OTRNPublisher, value: Boolean) {
         view.setScalableScreenshare(value)
+    }
+
+    @ReactProp(name = "allowAudioCaptureWhileMuted")
+    override public fun setAllowAudioCaptureWhileMuted(view: OTRNPublisher, value: Boolean) {
+        view.setAllowAudioCaptureWhileMuted(value)
     }
 
     companion object {