+++ ios/OpentokReactNative.swift
@@ -177,7 +177,8 @@ import React
         // Bug in OT iOS SDK. This is set to false, but it should be true:
         sessionCapabilities["canSubscribe"] = true;
         sessionCapabilities["canForceMute"] = session.capabilities?.canForceMute;
-        resolve([sessionCapabilities]);
+        sessionCapabilities["canForceDisconnect"] = session.capabilities?.canForceDisconnect;
+        resolve(sessionCapabilities);
     }
 
     @objc public func reportIssue(
@@ -320,6 +321,29 @@ import React
         }
         var error: OTError?
         session.disableForceMute(&error)
+        if let error = error {
+            reject("event_failure", error.localizedDescription, nil)
+            return
+        }
+        resolve(true)
+    }
+
+    @objc public func forceDisconnect(
+        _ sessionId: String,
+        connectionId: String,
+        resolve: @escaping RCTPromiseResolveBlock,
+        reject: @escaping RCTPromiseRejectBlock
+    ) {
+        guard let session = OTRN.sharedState.sessions[sessionId] else {
+            reject("event_failure", "Session ID not found", nil)
+            return
+        }
+        guard let connection = OTRN.sharedState.connections[connectionId] else {
+            reject("ERROR", "Connection ID not found", nil)
+            return
+        }
+        var error: OTError?
+        session.forceDisconnect(connection, error: &error)
         if let error = error {
             reject("event_failure", error.localizedDescription, nil)
             return