+++ src/OTPublisher.js
@@ -205,10 +205,13 @@ OTPublisher.defaultProps = {
     name: '',
     publishCaptions: false,
     scalableScreenshare: false,
+    allowAudioCaptureWhileMuted: false,
     resolution: 'MEDIUM',
     videoTrack: true,
     videoSource: 'camera',
     videoContentHint: '',
+    maxVideoBitrate: 0,
+    videoBitratePreset: 'default',
   },
   style: {
     flex: 1,